week 4 day 1

+++ zee5appspringboot/src/main/java/com/zee/zee5app/repository/UserRepository.java
@@ -1,6 +1,7 @@
 package com.zee.zee5app.repository;
 
 import java.math.BigDecimal;
+import java.math.BigInteger;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -11,6 +12,6 @@ import com.zee.zee5app.dto.Register;
 public interface UserRepository extends JpaRepository<Register, String> {
 
 	//custom jpa method. We wont write any defn, just signature(declaration)
-	Boolean existsByEmailAndContactnumber(String email,BigDecimal contactnumber);
+	Boolean existsByEmailAndContactnumber(String email,BigInteger contactnumber);
 	//Boolean existsByContactnumber(BigDecimal contactnumber);
 }
